Rename refresher.type package
