+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/schema/refresher/SchemaRefresherFactory.java
@@ -19,13 +19,13 @@ package org.apache.shardingsphere.infra.metadata.schema.refresher;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.apache.shardingsphere.infra.metadata.schema.refresher.impl.CreateIndexStatementSchemaRefresher;
-import org.apache.shardingsphere.infra.metadata.schema.refresher.impl.CreateTableStatementSchemaRefresher;
-import org.apache.shardingsphere.infra.metadata.schema.refresher.impl.CreateViewStatementSchemaRefresher;
-import org.apache.shardingsphere.infra.metadata.schema.refresher.impl.DropIndexStatementSchemaRefresher;
-import org.apache.shardingsphere.infra.metadata.schema.refresher.impl.AlterTableStatementSchemaRefresher;
-import org.apache.shardingsphere.infra.metadata.schema.refresher.impl.DropTableStatementSchemaRefresher;
-import org.apache.shardingsphere.infra.metadata.schema.refresher.impl.DropViewStatementSchemaRefresher;
+import org.apache.shardingsphere.infra.metadata.schema.refresher.type.CreateIndexStatementSchemaRefresher;
+import org.apache.shardingsphere.infra.metadata.schema.refresher.type.CreateTableStatementSchemaRefresher;
+import org.apache.shardingsphere.infra.metadata.schema.refresher.type.CreateViewStatementSchemaRefresher;
+import org.apache.shardingsphere.infra.metadata.schema.refresher.type.DropIndexStatementSchemaRefresher;
+import org.apache.shardingsphere.infra.metadata.schema.refresher.type.AlterTableStatementSchemaRefresher;
+import org.apache.shardingsphere.infra.metadata.schema.refresher.type.DropTableStatementSchemaRefresher;
+import org.apache.shardingsphere.infra.metadata.schema.refresher.type.DropViewStatementSchemaRefresher;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.ddl.AlterTableStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.ddl.CreateIndexStatement;
