+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/schema/refresher/type/AlterTableStatementSchemaRefresher.java
@@ -15,25 +15,28 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.refresher.impl;
+package org.apache.shardingsphere.infra.metadata.schema.refresher.type;
 
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
 import org.apache.shardingsphere.infra.metadata.schema.ShardingSphereSchema;
 import org.apache.shardingsphere.infra.metadata.schema.refresher.SchemaRefresher;
 import org.apache.shardingsphere.infra.metadata.schema.refresher.TableMetaDataLoaderCallback;
-import org.apache.shardingsphere.sql.parser.sql.common.statement.ddl.DropTableStatement;
+import org.apache.shardingsphere.sql.parser.sql.common.statement.ddl.AlterTableStatement;
 
+import java.sql.SQLException;
 import java.util.Collection;
 
 /**
- * ShardingSphere schema refresher for drop table statement.
+ * ShardingSphere schema refresher for alter table statement.
  */
-public final class DropTableStatementSchemaRefresher implements SchemaRefresher<DropTableStatement> {
+public final class AlterTableStatementSchemaRefresher implements SchemaRefresher<AlterTableStatement> {
     
     @Override
     public void refresh(final ShardingSphereSchema schema, final DatabaseType databaseType, final Collection<String> routeDataSourceNames,
-                        final DropTableStatement sqlStatement, final TableMetaDataLoaderCallback callback) {
-        sqlStatement.getTables().forEach(each -> schema.remove(each.getTableName().getIdentifier().getValue()));
+                        final AlterTableStatement sqlStatement, final TableMetaDataLoaderCallback callback) throws SQLException {
+        String tableName = sqlStatement.getTable().getTableName().getIdentifier().getValue();
+        if (null != schema && schema.containsTable(tableName)) {
+            callback.load(tableName).ifPresent(tableMetaData -> schema.put(tableName, tableMetaData));
+        }
     }
-
 }
