+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/schema/refresher/type/CreateTableStatementSchemaRefresher.java
@@ -15,28 +15,34 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.refresher.impl;
+package org.apache.shardingsphere.infra.metadata.schema.refresher.type;
 
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
 import org.apache.shardingsphere.infra.metadata.schema.ShardingSphereSchema;
+import org.apache.shardingsphere.infra.metadata.schema.model.TableMetaData;
 import org.apache.shardingsphere.infra.metadata.schema.refresher.SchemaRefresher;
 import org.apache.shardingsphere.infra.metadata.schema.refresher.TableMetaDataLoaderCallback;
-import org.apache.shardingsphere.sql.parser.sql.common.statement.ddl.AlterTableStatement;
+import org.apache.shardingsphere.sql.parser.sql.common.statement.ddl.CreateTableStatement;
 
 import java.sql.SQLException;
 import java.util.Collection;
+import java.util.Optional;
 
 /**
- * ShardingSphere schema refresher for alter table statement.
+ * ShardingSphere schema refresher for create table statement.
  */
-public final class AlterTableStatementSchemaRefresher implements SchemaRefresher<AlterTableStatement> {
+public final class CreateTableStatementSchemaRefresher implements SchemaRefresher<CreateTableStatement> {
     
     @Override
     public void refresh(final ShardingSphereSchema schema, final DatabaseType databaseType, final Collection<String> routeDataSourceNames,
-                        final AlterTableStatement sqlStatement, final TableMetaDataLoaderCallback callback) throws SQLException {
+                        final CreateTableStatement sqlStatement, final TableMetaDataLoaderCallback callback) throws SQLException {
         String tableName = sqlStatement.getTable().getTableName().getIdentifier().getValue();
-        if (null != schema && schema.containsTable(tableName)) {
-            callback.load(tableName).ifPresent(tableMetaData -> schema.put(tableName, tableMetaData));
+        Optional<TableMetaData> tableMetaData = callback.load(tableName);
+        if (tableMetaData.isPresent()) {
+            schema.put(tableName, tableMetaData.get());
+        } else {
+            schema.put(tableName, new TableMetaData());
         }
+        schema.get(tableName).getAddressingDataSources().addAll(routeDataSourceNames);
     }
 }
